+++ .azure-pipelines/PipelineSteps/BatchGeneration/filter.ps1
@@ -15,7 +15,21 @@ $artifactsDir = Join-Path $RepoRoot 'artifacts'
 
 $changedModulesDict = @{}
 $changedSubModulesDict = @{}
-if ($env:RUN_TEST_ON_ALL_MODULES -eq "True") {
+if ($env:TEST_CHANGED_MODULES_ONLY -eq 'true') {
+    Write-Host "Run test on generated folder changed modules"
+    # Only generated folder change should trigger the test
+    for ($i = 0; $i -lt $ChangedFiles.Count; $i++) {
+        if ($ChangedFiles[$i] -match '^generated/([^/]+)/([^/]+\.Autorest)/') {
+            $moduleName = $Matches[1]
+            $subModuleName = $Matches[2]
+            $subModule = "$moduleName/$subModuleName"
+            
+            $changedModulesDict[$moduleName] = $true
+            $changedSubModulesDict[$subModule] = $true
+        }
+    }
+}
+else {
     Write-Host "Run test on all modules"
     $V4ModulesFile = Join-Path $artifactsDir "generationTargets.json"
     $V4ModuleMaps = Get-Content -Raw -Path $V4ModulesFile | ConvertFrom-Json
@@ -28,20 +42,6 @@ if ($env:RUN_TEST_ON_ALL_MODULES -eq "True") {
                 $changedModulesDict[$moduleName] = $true
                 $changedSubModulesDict[$subModule] = $true
             }
-        }
-    }
-}
-else {
-    Write-Host "Run test on generated folder changed modules"
-    # Only generated folder change should trigger the test 
-    for ($i = 0; $i -lt $ChangedFiles.Count; $i++) {
-        if ($ChangedFiles[$i] -match '^generated/([^/]+)/([^/]+\.autorest)/') {
-            $moduleName = $Matches[2]
-            $subModuleName = $Matches[3]
-            $subModule = "$moduleName/$subModuleName"
-            
-            $changedModulesDict[$moduleName] = $true
-            $changedSubModulesDict[$subModule] = $true
         }
     }
 }
@@ -62,6 +62,11 @@ foreach ($subModule in $changedSubModules) {
 }
 Write-Host "##[endgroup]"
 Write-Host
+
+$changedModulesRecordFile = Join-Path $artifactsDir 'filteredChangedModules.txt'
+$changedModules | Set-Content -Path $changedModulesRecordFile -Encoding UTF8
+$changedSubModulesRecordFile = Join-Path $artifactsDir 'filteredChangedSubModules.txt'
+$changedSubModules | Set-Content -Path $changedSubModulesRecordFile -Encoding UTF8
 
 $groupedBuildModules = Group-Modules -Modules $changedModules -MaxParallelJobs $MaxParallelBuildJobs
 Write-Matrix -GroupedModules $groupedBuildModules -VariableName 'buildTargets' -RepoRoot $RepoRoot