+++ src/DataFactory/DataFactoryV2/Models/DataFactoryClient.IntegrationRuntimes.cs
@@ -484,7 +484,7 @@ namespace Microsoft.Azure.Commands.DataFactoryV2
                 }
             }
 
-            // Don't support get status for legacy integraiton runtime.
+            // Don't support get status for legacy integration runtime.
             throw new PSInvalidOperationException("This type of integration runtime is not supported by this version powershell cmdlets.");
         }