+++ src/Migrate/Migrate.Autorest/custom/New-AzMigrateServerReplication.ps1
@@ -197,7 +197,7 @@ function New-AzMigrateServerReplication {
         [Parameter(ParameterSetName = 'ByInputObjectDefaultUser')]
         [Microsoft.Azure.PowerShell.Cmdlets.Migrate.Category('Path')]
         [System.String]
-        # Specifies the disk encyption set to be used.
+        # Specifies the disk encryption set to be used.
         ${DiskEncryptionSetID},
     
         [Parameter()]
@@ -471,7 +471,8 @@ function New-AzMigrateServerReplication {
             $null = $PSBoundParameters.Remove('ResourceName')
             $null = $PSBoundParameters.Remove('SubscriptionId')
             $null = $PSBoundParameters.Add('Location', $TargetRegion)
-            $allAvailableSkus = Get-AzVMSize @PSBoundParameters -ErrorVariable notPresent -ErrorAction SilentlyContinue
+            #Get-AzVMSku -Location is deprecated, replicate using Get-AzComputeResourceSKU and a where clause
+            $allAvailableSkus = Get-AzComputeResourceSKU @PSBoundParameters| Where-Object { $_.ResourceType.Contains("virtualMachines") }
             if ($null -ne $allAvailableSkus) {
                 $matchingComputeSku = $allAvailableSkus | Where-Object { $_.Name -eq $TargetVMSize }
                 if ($null -ne $matchingComputeSku) {
@@ -720,7 +721,7 @@ public static int hashForArtifact(String artifact)
             }
 
             if (-not $HasTargetAVZone -and $disk.DiskType -eq "PremiumV2_LRS") {
-                throw "Premium SSD V2 disk can only be attached to zonal VMs."
+                throw "The Availability zone has not been provided for the Premium SSD V2 disk."
             }
 
             $res = $uniqueDiskUuids.Add($disk.DiskId)